+++ frontend/src/App.tsx
@@ -35,3 +35,4 @@ export default function App() {
     </Router>
   );
 }
+
