+++ frontend/src/components/common/Footer.tsx
@@ -1,12 +1,10 @@
 // src/components/common/Footer.tsx
 export function Footer() {
   return (
-    <footer id="contacto" className="bg-primary text-primary-foreground text-center py-8 px-6 mt-12">
-      <h4 className="text-xl font-semibold mb-2">Contáctanos</h4>
-      <p>Email: contacto@transmovil.com | Tel: +591 70000000</p>
-      <p className="mt-4 text-sm">
-        © 2025 TransMovil - Todos los derechos reservados
-      </p>
-    </footer>
+    <footer id="contacto" className="bg-blue-700 text-white text-center py-8 px-6 mt-12">
+        <h4 className="text-xl font-semibold mb-2">Contáctanos</h4>
+        <p>Email: contacto@transmovil.com | Tel: +591 70000000</p>
+        <p className="mt-4 text-sm">© 2025 TransMovil - Todos los derechos reservados</p>
+      </footer>
   );
 }
