+++ frontend/src/components/common/Navbar.tsx
@@ -1,24 +1,20 @@
 import { Link } from 'react-router-dom';
+import { Button } from "@/components/ui/button";
+import { Bus, Ticket, Package, Map, Users } from "lucide-react";
 
 const Navbar = () => {
   return (
-    <nav className="bg-white shadow-sm border-b border-gray-200">
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-        <div className="flex justify-between items-center h-16">
-          {/* Tu logo/marca existente */}
-          <div className="flex items-center">
-            <h1 className="text-2xl font-bold text-gray-900">
-              🚌 Transporte-SI2
-            </h1>
-          </div>
-          
-          {/* Tu menú existente */}
-          <div className="hidden md:flex items-center space-x-8">
-            {/* Tus enlaces existentes */}
-          </div>
-
-          {/* AGREGAR ESTOS BOTONES */}
-          <div className="flex items-center space-x-4">
+    <header className="flex justify-between items-center px-10 py-6 shadow-md bg-white sticky top-0 z-50">
+        <h1 className="text-2xl font-bold text-blue-700 flex items-center gap-2">
+          <Bus className="w-7 h-7" /> Transporte
+        </h1>
+        <nav className="space-x-6 hidden md:flex">
+          <a href="#servicios" className="hover:text-blue-600">Servicios</a>
+          <a href="#rutas" className="hover:text-blue-600">Rutas</a>
+          <a href="#choferes" className="hover:text-blue-600">Choferes</a>
+          <a href="#contacto" className="hover:text-blue-600">Contacto</a>
+        </nav>
+        <div className="flex items-center space-x-4">
             <Link 
               to="/login"
               className="px-4 py-2 text-gray-600 hover:text-gray-900 border border-gray-300 rounded-lg hover:border-gray-400 transition-colors"
@@ -29,13 +25,14 @@ const Navbar = () => {
               to="/register"
               className="px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors"
             >
-              Registrar
+              Crear Cuenta
             </Link>
           </div>
-        </div>
-      </div>
-    </nav>
+      </header>
+    
   );
 };
 
 export default Navbar;
+
+
